Move TrackTypeMustBeMadeInvisible() to Legacy.cpp, add MT
