+++ src/openrct2/park/Legacy.h
@@ -11,6 +11,7 @@
 
 #include "../object/Object.h"
 
+#include <cstdint>
 #include <optional>
 #include <string_view>
 
@@ -20,6 +21,8 @@ namespace OpenRCT2::RCT2
 }
 struct ObjectEntryDescriptor;
 class ObjectList;
+using ride_type_t = uint16_t;
+using track_type_t = uint16_t;
 
 std::string_view MapToNewObjectIdentifier(std::string_view s);
 std::optional<std::string_view> GetDATPathName(std::string_view newPathName);
@@ -28,3 +31,16 @@ void UpdateFootpathsFromMapping(
     ObjectEntryIndex* pathToSurfaceMap, ObjectEntryIndex* pathToQueueSurfaceMap, ObjectEntryIndex* pathToRailingsMap,
     ObjectList& requiredObjects, ObjectEntryIndex& surfaceCount, ObjectEntryIndex& railingCount, ObjectEntryIndex entryIndex,
     const OpenRCT2::RCT2::FootpathMapping* footpathMapping);
+
+/**
+ * If new pieces get added to existing ride types, this could cause existing parks to change appearance,
+ * since the formerly unrendered pieces were not explicitly set invisible.
+ * To avoid this, this function will return true if the piece is question was added after the park was created,
+ * so that import code can properly set the visibility.
+ *
+ * @param rideType The OpenRCT2 ride type
+ * @param trackType The OpenRCT2 track type
+ * @param parkFileVersion The current park file version. Pass -1 when converting S4 or S6.
+ * @return
+ */
+bool TrackTypeMustBeMadeInvisible(ride_type_t rideType, track_type_t trackType, int32_t parkFileVersion = -1);
