+++ src/openrct2/rct2/S6Importer.cpp
@@ -49,6 +49,7 @@
 #include "../object/ObjectManager.h"
 #include "../object/ObjectRepository.h"
 #include "../object/WallSceneryEntry.h"
+#include "../park/Legacy.h"
 #include "../peep/RideUseSystem.h"
 #include "../rct12/CSStringConverter.h"
 #include "../rct12/EntryList.h"
